# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kabianga ARG Portal API",
        "version": "1.0.0",
        "description": "Research grants management API"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": ["Users"],
                "summary": "List all users",
                "security": [{"bearerAuth": []}],
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                }
            }
        },
...
        "/api/v1/permissions": {
            "get": {
                "tags": ["Permissions"],
                "summary": "List all permissions",
                "security": [{"bearerAuth": []}],
                "responses": {
                    "200": {
                        "description": "List of permissions"
                    }
                }
            }
        },
...
        "/api/v1/users/search": {
            "get": {
                "tags": ["Users"],
                "summary": "Search users",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {"type": "string"}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                }
            }
        },
...
    }
}